query_cli/requirements3.txt
pytest
click
requests
boto
boto3
apache-ranger
snowflake-connector-python
